requirements.txt
annotated-types==0.7.0
anthropic==0.50.0
anyio==4.9.0
backoff==2.2.1
beautifulsoup4==4.13.4
boto3==1.38.9
botocore==1.38.9
browser-use==0.1.40
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cython==3.0.12
defusedxml==0.7.1
distro==1.9.0
faiss-cpu==1.11.0
filetype==1.2.0
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-auth==2.40.0
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jiter==0.9.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.22
langchain-anthropic==0.3.3
langchain-aws==0.2.19
langchain-core==0.3.49
langchain-deepseek==0.1.3
langchain-google-genai==2.1.2
langchain-ollama==0.2.2
langchain-openai==0.3.1
langchain-text-splitters==0.3.7
langsmith==0.3.42
linkify-it-py==2.0.3
markdown-it-py==3.0.0
markdownify==0.14.1
mdit-py-plugins==0.4.2
mdurl==0.1.2
mem0ai==0.1.93
monotonic==1.6
numpy==1.26.4
ollama==0.4.8
openai==1.77.0
orjson==3.10.18
packaging==24.2
patchright==1.52.0
platformdirs==4.3.7
playwright==1.52.0
portalocker==2.10.1
posthog==3.25.0
proto-plus==1.26.1
protobuf==6.30.2
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.10.6
pydantic-core==2.27.2
pyee==13.0.0
pygments==2.19.1
pyobjc-core==11.0
pyobjc-framework-cocoa==11.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
pyyaml==6.0.2
qdrant-client==1.14.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
s3transfer==0.12.0
screeninfo==0.8.1
setuptools==80.3.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.40
tenacity==9.1.2
textual==3.2.0
tiktoken==0.9.0
tqdm==4.67.1
typing-extensions==4.13.2
uc-micro-py==1.0.3
urllib3==2.4.0
zstandard==0.23.0
